+++ src/app/SBE/details/template/layouts/Actions.js
